requirements.txt
Django==1.6.2
South==0.8.4
gevent==1.0
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.2
wsgiref==0.1.2
